Fix for crash when hot_to_r on empty table
When you applyed the function hto_to_r on an empty data table
R would give an error as you try to create a matrix with NULL as
your data.  I have applyed a check to insure that there is data
to convert into a table, and if not an empty table will safly be
created.