Modify afterChange to work with column sort
I've modified the afterChangeCallback function so that it will work more effectively with sorted columns, or at least be easier to parse. The prior code would return a set of two lists per change, one with the row number and one with everything else. If columns were sorted and multiple values were changed, then it would only return one copy of this list. The new code makes it so that changes to sorted columns are treated the same way as everything else (namely, that they would return a list of four items per change).