fix hot_validate_numeric validator (#277)
As reported in issue #270 there was an issue with the numeric validator when a string is entered that begins with a digit but contains non-digits later.  The parseFloat function was the issue due to its robust design. I replaced it with the regexp approach used in handsontable's standard numeric validator. After that is the code that extends the validator functionality.